requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-redis==5.4.0
djangorestframework==3.16.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
marshmallow==4.0.0
prompt_toolkit==3.0.51
PyConfigure==0.5.9
PyStaticConfiguration==0.11.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
simplekv==0.13.0
six==1.17.0
sqlparse==0.5.3
structlog==25.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
ussd_airflow_engine @ git+https://github.com/jayskar/ussd_engine.git@0263868a59578483d442e1d26aa4008507ae604e
vine==5.1.0
wcwidth==0.2.13
